Added Template usage options to Template Edit Page
"Namespace Usage" shows all Assets that use the given template
namespace in their definition. The results can be expanded
to show definition properties.

"Instance Usage" shows all Asset instances that are currently
using the given template. Other templates in the same namespace
can be displayed, and the template can be replaced globally
with another template (to make it easy to decommission a template)