RFE #10741 - Account Modules over SSL/HTTPS
It is useful to be able to turn on SSL/HTTPS for Account Modules
in the same way that you can toggle "Encrypt Page" on a per-Asset
basis.

For example, the User account module that has the change password
screen is one page that you probably want to serve over SSL.

This commit uses a new per-account module "ssl" option in the site
config. When enabled, it causes the account module to force
delivery of the page over HTTPS (basically the same way Asset does
it).